修正：JS/フロントエンドスタックの配置
