+++ src/main/resources/static/assets/codeblock/codeblock.js
@@ -1,0 +1,8 @@
+function dt(E){return E&&E.__esModule&&Object.prototype.hasOwnProperty.call(E,"default")?E.default:E}var We,ut;function Qt(){if(ut)return We;ut=1;function E(e){return e instanceof Map?e.clear=e.delete=e.set=function(){throw new Error("map is read-only")}:e instanceof Set&&(e.add=e.clear=e.delete=function(){throw new Error("set is read-only")}),Object.freeze(e),Object.getOwnPropertyNames(e).forEach(t=>{const r=e[t],p=typeof r;(p==="object"||p==="function")&&!Object.isFrozen(r)&&E(r)}),e}class Y{constructor(t){t.data===void 0&&(t.data={}),this.data=t.data,this.isMatchIgnored=!1}ignoreMatch(){this.isMatchIgnored=!0}}function j(e){return e.replace(/&/g,"&amp;").replace(/</g,"&lt;").replace(/>/g,"&gt;").replace(/"/g,"&quot;").replace(/'/g,"&#x27;")}function U(e,...t){const r=Object.create(null);for(const p in e)r[p]=e[p];return t.forEach(function(p){for(const I in p)r[I]=p[I]}),r}const q="</span>",S=e=>!!e.scope,w=(e,{prefix:t})=>{if(e.startsWith("language:"))return e.replace("language:","language-");if(e.includes(".")){const r=e.split(".");return[`${t}${r.shift()}`,...r.map((p,I)=>`${p}${"_".repeat(I+1)}`)].join(" ")}return`${t}${e}`};class o{constructor(t,r){this.buffer="",this.classPrefix=r.classPrefix,t.walk(this)}addText(t){this.buffer+=j(t)}openNode(t){if(!S(t))return;const r=w(t.scope,{prefix:this.classPrefix});this.span(r)}closeNode(t){S(t)&&(this.buffer+=q)}value(){return this.buffer}span(t){this.buffer+=`<span class="${t}">`}}const _=(e={})=>{const t={children:[]};return Object.assign(t,e),t};class v{constructor(){this.rootNode=_(),this.stack=[this.rootNode]}get top(){return this.stack[this.stack.length-1]}get root(){return this.rootNode}add(t){this.top.children.push(t)}openNode(t){const r=_({scope:t});this.add(r),this.stack.push(r)}closeNode(){if(this.stack.length>1)return this.stack.pop()}closeAllNodes(){for(;this.closeNode(););}toJSON(){return JSON.stringify(this.rootNode,null,4)}walk(t){return this.constructor._walk(t,this.rootNode)}static _walk(t,r){return typeof r=="string"?t.addText(r):r.children&&(t.openNode(r),r.children.forEach(p=>this._walk(t,p)),t.closeNode(r)),t}static _collapse(t){typeof t!="string"&&t.children&&(t.children.every(r=>typeof r=="string")?t.children=[t.children.join("")]:t.children.forEach(r=>{v._collapse(r)}))}}class x extends v{constructor(t){super(),this.options=t}addText(t){t!==""&&this.add(t)}startScope(t){this.openNode(t)}endScope(){this.closeNode()}__addSublanguage(t,r){const p=t.root;r&&(p.scope=`language:${r}`),this.add(p)}toHTML(){return new o(this,this.options).value()}finalize(){return this.closeAllNodes(),!0}}function C(e){return e?typeof e=="string"?e:e.source:null}function L(e){return R("(?=",e,")")}function F(e){return R("(?:",e,")*")}function N(e){return R("(?:",e,")?")}function R(...e){return e.map(r=>C(r)).join("")}function T(e){const t=e[e.length-1];return typeof t=="object"&&t.constructor===Object?(e.splice(e.length-1,1),t):{}}function G(...e){return"("+(T(e).capture?"":"?:")+e.map(p=>C(p)).join("|")+")"}function oe(e){return new RegExp(e.toString()+"|").exec("").length-1}function _e(e,t){const r=e&&e.exec(t);return r&&r.index===0}const ce=/\[(?:[^\\\]]|\\.)*\]|\(\??|\\([1-9][0-9]*)|\\./;function K(e,{joinWith:t}){let r=0;return e.map(p=>{r+=1;const I=r;let B=C(p),a="";for(;B.length>0;){const s=ce.exec(B);if(!s){a+=B;break}a+=B.substring(0,s.index),B=B.substring(s.index+s[0].length),s[0][0]==="\\"&&s[1]?a+="\\"+String(Number(s[1])+I):(a+=s[0],s[0]==="("&&r++)}return a}).map(p=>`(${p})`).join(t)}const ne=/\b\B/,ge="[a-zA-Z]\\w*",Z="[a-zA-Z_]\\w*",J="\\b\\d+(\\.\\d+)?",be="(-?)(\\b0[xX][a-fA-F0-9]+|(\\b\\d+(\\.\\d*)?|\\.\\d+)([eE][-+]?\\d+)?)",xe="\\b(0b[01]+)",De="!|!=|!==|%|%=|&|&&|&=|\\*|\\*=|\\+|\\+=|,|-|-=|/=|/|:|;|<<|<<=|<=|<|===|==|=|>>>=|>>=|>=|>>>|>>|>|\\?|\\[|\\{|\\(|\\^|\\^=|\\||\\|=|\\|\\||~",ye=(e={})=>{const t=/^#![ ]*\//;return e.binary&&(e.begin=R(t,/.*\b/,e.binary,/\b.*/)),U({scope:"meta",begin:t,end:/$/,relevance:0,"on:begin":(r,p)=>{r.index!==0&&p.ignoreMatch()}},e)},le={begin:"\\\\[\\s\\S]",relevance:0},Le={scope:"string",begin:"'",end:"'",illegal:"\\n",contains:[le]},Ie={scope:"string",begin:'"',end:'"',illegal:"\\n",contains:[le]},Be={begin:/\b(a|an|the|are|I'm|isn't|don't|doesn't|won't|but|just|should|pretty|simply|enough|gonna|going|wtf|so|such|will|you|your|they|like|more)\b/},re=function(e,t,r={}){const p=U({scope:"comment",begin:e,end:t,contains:[]},r);p.contains.push({scope:"doctag",begin:"[ ]*(?=(TODO|FIXME|NOTE|BUG|OPTIMIZE|HACK|XXX):)",end:/(TODO|FIXME|NOTE|BUG|OPTIMIZE|HACK|XXX):/,excludeBegin:!0,relevance:0});const I=G("I","a","is","so","us","to","at","if","in","it","on",/[A-Za-z]+['](d|ve|re|ll|t|s|n)/,/[A-Za-z]+[-][a-z]+/,/[A-Za-z][a-z]{2,}/);return p.contains.push({begin:R(/[ ]+/,"(",I,/[.]?[:]?([.][ ]|[ ])/,"){3}")}),p},Ee=re("//","$"),Pe=re("/\\*","\\*/"),je=re("#","$"),y={scope:"number",begin:J,relevance:0},u={scope:"number",begin:be,relevance:0},c={scope:"number",begin:xe,relevance:0},d={scope:"regexp",begin:/\/(?=[^/\n]*\/)/,end:/\/[gimuy]*/,contains:[le,{begin:/\[/,end:/\]/,relevance:0,contains:[le]}]},b={scope:"title",begin:ge,relevance:0},O={scope:"title",begin:Z,relevance:0},X={begin:"\\.\\s*"+Z,relevance:0};var me=Object.freeze({__proto__:null,APOS_STRING_MODE:Le,BACKSLASH_ESCAPE:le,BINARY_NUMBER_MODE:c,BINARY_NUMBER_RE:xe,COMMENT:re,C_BLOCK_COMMENT_MODE:Pe,C_LINE_COMMENT_MODE:Ee,C_NUMBER_MODE:u,C_NUMBER_RE:be,END_SAME_AS_BEGIN:function(e){return Object.assign(e,{"on:begin":(t,r)=>{r.data._beginMatch=t[1]},"on:end":(t,r)=>{r.data._beginMatch!==t[1]&&r.ignoreMatch()}})},HASH_COMMENT_MODE:je,IDENT_RE:ge,MATCH_NOTHING_RE:ne,METHOD_GUARD:X,NUMBER_MODE:y,NUMBER_RE:J,PHRASAL_WORDS_MODE:Be,QUOTE_STRING_MODE:Ie,REGEXP_MODE:d,RE_STARTERS_RE:De,SHEBANG:ye,TITLE_MODE:b,UNDERSCORE_IDENT_RE:Z,UNDERSCORE_TITLE_MODE:O});function ht(e,t){e.input[e.index-1]==="."&&t.ignoreMatch()}function _t(e,t){e.className!==void 0&&(e.scope=e.className,delete e.className)}function bt(e,t){t&&e.beginKeywords&&(e.begin="\\b("+e.beginKeywords.split(" ").join("|")+")(?!\\.)(?=\\b|\\s)",e.__beforeBegin=ht,e.keywords=e.keywords||e.beginKeywords,delete e.beginKeywords,e.relevance===void 0&&(e.relevance=0))}function yt(e,t){Array.isArray(e.illegal)&&(e.illegal=G(...e.illegal))}function Et(e,t){if(e.match){if(e.begin||e.end)throw new Error("begin & end are not supported with match");e.begin=e.match,delete e.match}}function vt(e,t){e.relevance===void 0&&(e.relevance=1)}const mt=(e,t)=>{if(!e.beforeMatch)return;if(e.starts)throw new Error("beforeMatch cannot be used with starts");const r=Object.assign({},e);Object.keys(e).forEach(p=>{delete e[p]}),e.keywords=r.keywords,e.begin=R(r.beforeMatch,L(r.begin)),e.starts={relevance:0,contains:[Object.assign(r,{endsParent:!0})]},e.relevance=0,delete r.beforeMatch},wt=["of","and","for","in","not","or","if","then","parent","list","value"],xt="keyword";function Ye(e,t,r=xt){const p=Object.create(null);return typeof e=="string"?I(r,e.split(" ")):Array.isArray(e)?I(r,e):Object.keys(e).forEach(function(B){Object.assign(p,Ye(e[B],t,B))}),p;function I(B,a){t&&(a=a.map(s=>s.toLowerCase())),a.forEach(function(s){const g=s.split("|");p[g[0]]=[B,Ot(g[0],g[1])]})}}function Ot(e,t){return t?Number(t):Mt(e)?0:1}function Mt(e){return wt.includes(e.toLowerCase())}const Xe={},ue=e=>{console.error(e)},Ve=(e,...t)=>{console.log(`WARN: ${e}`,...t)},de=(e,t)=>{Xe[`${e}/${t}`]||(console.log(`Deprecated as of ${e}. ${t}`),Xe[`${e}/${t}`]=!0)},Oe=new Error;function Ze(e,t,{key:r}){let p=0;const I=e[r],B={},a={};for(let s=1;s<=t.length;s++)a[s+p]=I[s],B[s+p]=!0,p+=oe(t[s-1]);e[r]=a,e[r]._emit=B,e[r]._multi=!0}function St(e){if(Array.isArray(e.begin)){if(e.skip||e.excludeBegin||e.returnBegin)throw ue("skip, excludeBegin, returnBegin not compatible with beginScope: {}"),Oe;if(typeof e.beginScope!="object"||e.beginScope===null)throw ue("beginScope must be object"),Oe;Ze(e,e.begin,{key:"beginScope"}),e.begin=K(e.begin,{joinWith:""})}}function Nt(e){if(Array.isArray(e.end)){if(e.skip||e.excludeEnd||e.returnEnd)throw ue("skip, excludeEnd, returnEnd not compatible with endScope: {}"),Oe;if(typeof e.endScope!="object"||e.endScope===null)throw ue("endScope must be object"),Oe;Ze(e,e.end,{key:"endScope"}),e.end=K(e.end,{joinWith:""})}}function Rt(e){e.scope&&typeof e.scope=="object"&&e.scope!==null&&(e.beginScope=e.scope,delete e.scope)}function Tt(e){Rt(e),typeof e.beginScope=="string"&&(e.beginScope={_wrap:e.beginScope}),typeof e.endScope=="string"&&(e.endScope={_wrap:e.endScope}),St(e),Nt(e)}function At(e){function t(a,s){return new RegExp(C(a),"m"+(e.case_insensitive?"i":"")+(e.unicodeRegex?"u":"")+(s?"g":""))}class r{constructor(){this.matchIndexes={},this.regexes=[],this.matchAt=1,this.position=0}addRule(s,g){g.position=this.position++,this.matchIndexes[this.matchAt]=g,this.regexes.push([g,s]),this.matchAt+=oe(s)+1}compile(){this.regexes.length===0&&(this.exec=()=>null);const s=this.regexes.map(g=>g[1]);this.matcherRe=t(K(s,{joinWith:"|"}),!0),this.lastIndex=0}exec(s){this.matcherRe.lastIndex=this.lastIndex;const g=this.matcherRe.exec(s);if(!g)return null;const $=g.findIndex((we,Ue)=>Ue>0&&we!==void 0),P=this.matchIndexes[$];return g.splice(0,$),Object.assign(g,P)}}class p{constructor(){this.rules=[],this.multiRegexes=[],this.count=0,this.lastIndex=0,this.regexIndex=0}getMatcher(s){if(this.multiRegexes[s])return this.multiRegexes[s];const g=new r;return this.rules.slice(s).forEach(([$,P])=>g.addRule($,P)),g.compile(),this.multiRegexes[s]=g,g}resumingScanAtSamePosition(){return this.regexIndex!==0}considerAll(){this.regexIndex=0}addRule(s,g){this.rules.push([s,g]),g.type==="begin"&&this.count++}exec(s){const g=this.getMatcher(this.regexIndex);g.lastIndex=this.lastIndex;let $=g.exec(s);if(this.resumingScanAtSamePosition()&&!($&&$.index===this.lastIndex)){const P=this.getMatcher(0);P.lastIndex=this.lastIndex+1,$=P.exec(s)}return $&&(this.regexIndex+=$.position+1,this.regexIndex===this.count&&this.considerAll()),$}}function I(a){const s=new p;return a.contains.forEach(g=>s.addRule(g.begin,{rule:g,type:"begin"})),a.terminatorEnd&&s.addRule(a.terminatorEnd,{type:"end"}),a.illegal&&s.addRule(a.illegal,{type:"illegal"}),s}function B(a,s){const g=a;if(a.isCompiled)return g;[_t,Et,Tt,mt].forEach(P=>P(a,s)),e.compilerExtensions.forEach(P=>P(a,s)),a.__beforeBegin=null,[bt,yt,vt].forEach(P=>P(a,s)),a.isCompiled=!0;let $=null;return typeof a.keywords=="object"&&a.keywords.$pattern&&(a.keywords=Object.assign({},a.keywords),$=a.keywords.$pattern,delete a.keywords.$pattern),$=$||/\w+/,a.keywords&&(a.keywords=Ye(a.keywords,e.case_insensitive)),g.keywordPatternRe=t($,!0),s&&(a.begin||(a.begin=/\B|\b/),g.beginRe=t(g.begin),!a.end&&!a.endsWithParent&&(a.end=/\B|\b/),a.end&&(g.endRe=t(g.end)),g.terminatorEnd=C(g.end)||"",a.endsWithParent&&s.terminatorEnd&&(g.terminatorEnd+=(a.end?"|":"")+s.terminatorEnd)),a.illegal&&(g.illegalRe=t(a.illegal)),a.contains||(a.contains=[]),a.contains=[].concat(...a.contains.map(function(P){return Ct(P==="self"?a:P)})),a.contains.forEach(function(P){B(P,g)}),a.starts&&B(a.starts,s),g.matcher=I(g),g}if(e.compilerExtensions||(e.compilerExtensions=[]),e.contains&&e.contains.includes("self"))throw new Error("ERR: contains `self` is not supported at the top-level of a language.  See documentation.");return e.classNameAliases=U(e.classNameAliases||{}),B(e)}function Je(e){return e?e.endsWithParent||Je(e.starts):!1}function Ct(e){return e.variants&&!e.cachedVariants&&(e.cachedVariants=e.variants.map(function(t){return U(e,{variants:null},t)})),e.cachedVariants?e.cachedVariants:Je(e)?U(e,{starts:e.starts?U(e.starts):null}):Object.isFrozen(e)?U(e):e}var kt="11.11.1";class Dt extends Error{constructor(t,r){super(t),this.name="HTMLInjectionError",this.html=r}}const He=j,Qe=U,et=Symbol("nomatch"),Lt=7,tt=function(e){const t=Object.create(null),r=Object.create(null),p=[];let I=!0;const B="Could not find the language '{}', did you forget to load/include a language module?",a={disableAutodetect:!0,name:"Plain text",contains:[]};let s={ignoreUnescapedHTML:!1,throwUnescapedHTML:!1,noHighlightRe:/^(no-?highlight)$/i,languageDetectRe:/\blang(?:uage)?-([\w-]+)\b/i,classPrefix:"hljs-",cssSelector:"pre code",languages:null,__emitter:x};function g(n){return s.noHighlightRe.test(n)}function $(n){let f=n.className+" ";f+=n.parentNode?n.parentNode.className:"";const M=s.languageDetectRe.exec(f);if(M){const k=se(M[1]);return k||(Ve(B.replace("{}",M[1])),Ve("Falling back to no-highlight mode for this block.",n)),k?M[1]:"no-highlight"}return f.split(/\s+/).find(k=>g(k)||se(k))}function P(n,f,M){let k="",H="";typeof f=="object"?(k=n,M=f.ignoreIllegals,H=f.language):(de("10.7.0","highlight(lang, code, ...args) has been deprecated."),de("10.7.0",`Please use highlight(code, options) instead.
+https://github.com/highlightjs/highlight.js/issues/2277`),H=n,k=f),M===void 0&&(M=!0);const Q={code:k,language:H};Se("before:highlight",Q);const ae=Q.result?Q.result:we(Q.language,Q.code,M);return ae.code=Q.code,Se("after:highlight",ae),ae}function we(n,f,M,k){const H=Object.create(null);function Q(i,l){return i.keywords[l]}function ae(){if(!h.keywords){z.addText(D);return}let i=0;h.keywordPatternRe.lastIndex=0;let l=h.keywordPatternRe.exec(D),m="";for(;l;){m+=D.substring(i,l.index);const A=te.case_insensitive?l[0].toLowerCase():l[0],W=Q(h,A);if(W){const[ie,Zt]=W;if(z.addText(m),m="",H[A]=(H[A]||0)+1,H[A]<=Lt&&(Te+=Zt),ie.startsWith("_"))m+=l[0];else{const Jt=te.classNameAliases[ie]||ie;ee(l[0],Jt)}}else m+=l[0];i=h.keywordPatternRe.lastIndex,l=h.keywordPatternRe.exec(D)}m+=D.substring(i),z.addText(m)}function Ne(){if(D==="")return;let i=null;if(typeof h.subLanguage=="string"){if(!t[h.subLanguage]){z.addText(D);return}i=we(h.subLanguage,D,!0,lt[h.subLanguage]),lt[h.subLanguage]=i._top}else i=$e(D,h.subLanguage.length?h.subLanguage:null);h.relevance>0&&(Te+=i.relevance),z.__addSublanguage(i._emitter,i.language)}function V(){h.subLanguage!=null?Ne():ae(),D=""}function ee(i,l){i!==""&&(z.startScope(l),z.addText(i),z.endScope())}function ot(i,l){let m=1;const A=l.length-1;for(;m<=A;){if(!i._emit[m]){m++;continue}const W=te.classNameAliases[i[m]]||i[m],ie=l[m];W?ee(ie,W):(D=ie,ae(),D=""),m++}}function st(i,l){return i.scope&&typeof i.scope=="string"&&z.openNode(te.classNameAliases[i.scope]||i.scope),i.beginScope&&(i.beginScope._wrap?(ee(D,te.classNameAliases[i.beginScope._wrap]||i.beginScope._wrap),D=""):i.beginScope._multi&&(ot(i.beginScope,l),D="")),h=Object.create(i,{parent:{value:h}}),h}function at(i,l,m){let A=_e(i.endRe,m);if(A){if(i["on:end"]){const W=new Y(i);i["on:end"](l,W),W.isMatchIgnored&&(A=!1)}if(A){for(;i.endsParent&&i.parent;)i=i.parent;return i}}if(i.endsWithParent)return at(i.parent,l,m)}function Wt(i){return h.matcher.regexIndex===0?(D+=i[0],1):(Ke=!0,0)}function qt(i){const l=i[0],m=i.rule,A=new Y(m),W=[m.__beforeBegin,m["on:begin"]];for(const ie of W)if(ie&&(ie(i,A),A.isMatchIgnored))return Wt(l);return m.skip?D+=l:(m.excludeBegin&&(D+=l),V(),!m.returnBegin&&!m.excludeBegin&&(D=l)),st(m,i),m.returnBegin?0:l.length}function Yt(i){const l=i[0],m=f.substring(i.index),A=at(h,i,m);if(!A)return et;const W=h;h.endScope&&h.endScope._wrap?(V(),ee(l,h.endScope._wrap)):h.endScope&&h.endScope._multi?(V(),ot(h.endScope,i)):W.skip?D+=l:(W.returnEnd||W.excludeEnd||(D+=l),V(),W.excludeEnd&&(D=l));do h.scope&&z.closeNode(),!h.skip&&!h.subLanguage&&(Te+=h.relevance),h=h.parent;while(h!==A.parent);return A.starts&&st(A.starts,i),W.returnEnd?0:l.length}function Xt(){const i=[];for(let l=h;l!==te;l=l.parent)l.scope&&i.unshift(l.scope);i.forEach(l=>z.openNode(l))}let Re={};function ct(i,l){const m=l&&l[0];if(D+=i,m==null)return V(),0;if(Re.type==="begin"&&l.type==="end"&&Re.index===l.index&&m===""){if(D+=f.slice(l.index,l.index+1),!I){const A=new Error(`0 width match regex (${n})`);throw A.languageName=n,A.badRule=Re.rule,A}return 1}if(Re=l,l.type==="begin")return qt(l);if(l.type==="illegal"&&!M){const A=new Error('Illegal lexeme "'+m+'" for mode "'+(h.scope||"<unnamed>")+'"');throw A.mode=h,A}else if(l.type==="end"){const A=Yt(l);if(A!==et)return A}if(l.type==="illegal"&&m==="")return D+=`
+`,1;if(Ge>1e5&&Ge>l.index*3)throw new Error("potential infinite loop, way more iterations than matches");return D+=m,m.length}const te=se(n);if(!te)throw ue(B.replace("{}",n)),new Error('Unknown language: "'+n+'"');const Vt=At(te);let ze="",h=k||Vt;const lt={},z=new s.__emitter(s);Xt();let D="",Te=0,fe=0,Ge=0,Ke=!1;try{if(te.__emitTokens)te.__emitTokens(f,z);else{for(h.matcher.considerAll();;){Ge++,Ke?Ke=!1:h.matcher.considerAll(),h.matcher.lastIndex=fe;const i=h.matcher.exec(f);if(!i)break;const l=f.substring(fe,i.index),m=ct(l,i);fe=i.index+m}ct(f.substring(fe))}return z.finalize(),ze=z.toHTML(),{language:n,value:ze,relevance:Te,illegal:!1,_emitter:z,_top:h}}catch(i){if(i.message&&i.message.includes("Illegal"))return{language:n,value:He(f),illegal:!0,relevance:0,_illegalBy:{message:i.message,index:fe,context:f.slice(fe-100,fe+100),mode:i.mode,resultSoFar:ze},_emitter:z};if(I)return{language:n,value:He(f),illegal:!1,relevance:0,errorRaised:i,_emitter:z,_top:h};throw i}}function Ue(n){const f={value:He(n),illegal:!1,relevance:0,_top:a,_emitter:new s.__emitter(s)};return f._emitter.addText(n),f}function $e(n,f){f=f||s.languages||Object.keys(t);const M=Ue(n),k=f.filter(se).filter(it).map(V=>we(V,n,!1));k.unshift(M);const H=k.sort((V,ee)=>{if(V.relevance!==ee.relevance)return ee.relevance-V.relevance;if(V.language&&ee.language){if(se(V.language).supersetOf===ee.language)return 1;if(se(ee.language).supersetOf===V.language)return-1}return 0}),[Q,ae]=H,Ne=Q;return Ne.secondBest=ae,Ne}function It(n,f,M){const k=f&&r[f]||M;n.classList.add("hljs"),n.classList.add(`language-${k}`)}function Fe(n){let f=null;const M=$(n);if(g(M))return;if(Se("before:highlightElement",{el:n,language:M}),n.dataset.highlighted){console.log("Element previously highlighted. To highlight again, first unset `dataset.highlighted`.",n);return}if(n.children.length>0&&(s.ignoreUnescapedHTML||(console.warn("One of your code blocks includes unescaped HTML. This is a potentially serious security risk."),console.warn("https://github.com/highlightjs/highlight.js/wiki/security"),console.warn("The element with unescaped HTML:"),console.warn(n)),s.throwUnescapedHTML))throw new Dt("One of your code blocks includes unescaped HTML.",n.innerHTML);f=n;const k=f.textContent,H=M?P(k,{language:M,ignoreIllegals:!0}):$e(k);n.innerHTML=H.value,n.dataset.highlighted="yes",It(n,M,H.language),n.result={language:H.language,re:H.relevance,relevance:H.relevance},H.secondBest&&(n.secondBest={language:H.secondBest.language,relevance:H.secondBest.relevance}),Se("after:highlightElement",{el:n,result:H,text:k})}function Bt(n){s=Qe(s,n)}const Pt=()=>{Me(),de("10.6.0","initHighlighting() deprecated.  Use highlightAll() now.")};function jt(){Me(),de("10.6.0","initHighlightingOnLoad() deprecated.  Use highlightAll() now.")}let nt=!1;function Me(){function n(){Me()}if(document.readyState==="loading"){nt||window.addEventListener("DOMContentLoaded",n,!1),nt=!0;return}document.querySelectorAll(s.cssSelector).forEach(Fe)}function Ht(n,f){let M=null;try{M=f(e)}catch(k){if(ue("Language definition for '{}' could not be registered.".replace("{}",n)),I)ue(k);else throw k;M=a}M.name||(M.name=n),t[n]=M,M.rawDefinition=f.bind(null,e),M.aliases&&rt(M.aliases,{languageName:n})}function Ut(n){delete t[n];for(const f of Object.keys(r))r[f]===n&&delete r[f]}function $t(){return Object.keys(t)}function se(n){return n=(n||"").toLowerCase(),t[n]||t[r[n]]}function rt(n,{languageName:f}){typeof n=="string"&&(n=[n]),n.forEach(M=>{r[M.toLowerCase()]=f})}function it(n){const f=se(n);return f&&!f.disableAutodetect}function Ft(n){n["before:highlightBlock"]&&!n["before:highlightElement"]&&(n["before:highlightElement"]=f=>{n["before:highlightBlock"](Object.assign({block:f.el},f))}),n["after:highlightBlock"]&&!n["after:highlightElement"]&&(n["after:highlightElement"]=f=>{n["after:highlightBlock"](Object.assign({block:f.el},f))})}function zt(n){Ft(n),p.push(n)}function Gt(n){const f=p.indexOf(n);f!==-1&&p.splice(f,1)}function Se(n,f){const M=n;p.forEach(function(k){k[M]&&k[M](f)})}function Kt(n){return de("10.7.0","highlightBlock will be removed entirely in v12.0"),de("10.7.0","Please use highlightElement now."),Fe(n)}Object.assign(e,{highlight:P,highlightAuto:$e,highlightAll:Me,highlightElement:Fe,highlightBlock:Kt,configure:Bt,initHighlighting:Pt,initHighlightingOnLoad:jt,registerLanguage:Ht,unregisterLanguage:Ut,listLanguages:$t,getLanguage:se,registerAliases:rt,autoDetection:it,inherit:Qe,addPlugin:zt,removePlugin:Gt}),e.debugMode=function(){I=!1},e.safeMode=function(){I=!0},e.versionString=kt,e.regex={concat:R,lookahead:L,either:G,optional:N,anyNumberOfTimes:F};for(const n in me)typeof me[n]=="object"&&E(me[n]);return Object.assign(e,me),e},pe=tt({});return pe.newInstance=()=>tt({}),We=pe,pe.HighlightJS=pe,pe.default=pe,We}var en=Qt();const qe=dt(en);function tn(E){const Y=E.regex,j=E.COMMENT("--","$"),U={scope:"string",variants:[{begin:/'/,end:/'/,contains:[{match:/''/}]}]},q={begin:/"/,end:/"/,contains:[{match:/""/}]},S=["true","false","unknown"],w=["double precision","large object","with timezone","without timezone"],o=["bigint","binary","blob","boolean","char","character","clob","date","dec","decfloat","decimal","float","int","integer","interval","nchar","nclob","national","numeric","real","row","smallint","time","timestamp","varchar","varying","varbinary"],_=["add","asc","collation","desc","final","first","last","view"],v=["abs","acos","all","allocate","alter","and","any","are","array","array_agg","array_max_cardinality","as","asensitive","asin","asymmetric","at","atan","atomic","authorization","avg","begin","begin_frame","begin_partition","between","bigint","binary","blob","boolean","both","by","call","called","cardinality","cascaded","case","cast","ceil","ceiling","char","char_length","character","character_length","check","classifier","clob","close","coalesce","collate","collect","column","commit","condition","connect","constraint","contains","convert","copy","corr","corresponding","cos","cosh","count","covar_pop","covar_samp","create","cross","cube","cume_dist","current","current_catalog","current_date","current_default_transform_group","current_path","current_role","current_row","current_schema","current_time","current_timestamp","current_path","current_role","current_transform_group_for_type","current_user","cursor","cycle","date","day","deallocate","dec","decimal","decfloat","declare","default","define","delete","dense_rank","deref","describe","deterministic","disconnect","distinct","double","drop","dynamic","each","element","else","empty","end","end_frame","end_partition","end-exec","equals","escape","every","except","exec","execute","exists","exp","external","extract","false","fetch","filter","first_value","float","floor","for","foreign","frame_row","free","from","full","function","fusion","get","global","grant","group","grouping","groups","having","hold","hour","identity","in","indicator","initial","inner","inout","insensitive","insert","int","integer","intersect","intersection","interval","into","is","join","json_array","json_arrayagg","json_exists","json_object","json_objectagg","json_query","json_table","json_table_primitive","json_value","lag","language","large","last_value","lateral","lead","leading","left","like","like_regex","listagg","ln","local","localtime","localtimestamp","log","log10","lower","match","match_number","match_recognize","matches","max","member","merge","method","min","minute","mod","modifies","module","month","multiset","national","natural","nchar","nclob","new","no","none","normalize","not","nth_value","ntile","null","nullif","numeric","octet_length","occurrences_regex","of","offset","old","omit","on","one","only","open","or","order","out","outer","over","overlaps","overlay","parameter","partition","pattern","per","percent","percent_rank","percentile_cont","percentile_disc","period","portion","position","position_regex","power","precedes","precision","prepare","primary","procedure","ptf","range","rank","reads","real","recursive","ref","references","referencing","regr_avgx","regr_avgy","regr_count","regr_intercept","regr_r2","regr_slope","regr_sxx","regr_sxy","regr_syy","release","result","return","returns","revoke","right","rollback","rollup","row","row_number","rows","running","savepoint","scope","scroll","search","second","seek","select","sensitive","session_user","set","show","similar","sin","sinh","skip","smallint","some","specific","specifictype","sql","sqlexception","sqlstate","sqlwarning","sqrt","start","static","stddev_pop","stddev_samp","submultiset","subset","substring","substring_regex","succeeds","sum","symmetric","system","system_time","system_user","table","tablesample","tan","tanh","then","time","timestamp","timezone_hour","timezone_minute","to","trailing","translate","translate_regex","translation","treat","trigger","trim","trim_array","true","truncate","uescape","union","unique","unknown","unnest","update","upper","user","using","value","values","value_of","var_pop","var_samp","varbinary","varchar","varying","versioning","when","whenever","where","width_bucket","window","with","within","without","year"],x=["abs","acos","array_agg","asin","atan","avg","cast","ceil","ceiling","coalesce","corr","cos","cosh","count","covar_pop","covar_samp","cume_dist","dense_rank","deref","element","exp","extract","first_value","floor","json_array","json_arrayagg","json_exists","json_object","json_objectagg","json_query","json_table","json_table_primitive","json_value","lag","last_value","lead","listagg","ln","log","log10","lower","max","min","mod","nth_value","ntile","nullif","percent_rank","percentile_cont","percentile_disc","position","position_regex","power","rank","regr_avgx","regr_avgy","regr_count","regr_intercept","regr_r2","regr_slope","regr_sxx","regr_sxy","regr_syy","row_number","sin","sinh","sqrt","stddev_pop","stddev_samp","substring","substring_regex","sum","tan","tanh","translate","translate_regex","treat","trim","trim_array","unnest","upper","value_of","var_pop","var_samp","width_bucket"],C=["current_catalog","current_date","current_default_transform_group","current_path","current_role","current_schema","current_transform_group_for_type","current_user","session_user","system_time","system_user","current_time","localtime","current_timestamp","localtimestamp"],L=["create table","insert into","primary key","foreign key","not null","alter table","add constraint","grouping sets","on overflow","character set","respect nulls","ignore nulls","nulls first","nulls last","depth first","breadth first"],F=x,N=[...v,..._].filter(K=>!x.includes(K)),R={scope:"variable",match:/@[a-z0-9][a-z0-9_]*/},T={scope:"operator",match:/[-+*/=%^~]|&&?|\|\|?|!=?|<(?:=>?|<|>)?|>[>=]?/,relevance:0},G={match:Y.concat(/\b/,Y.either(...F),/\s*\(/),relevance:0,keywords:{built_in:F}};function oe(K){return Y.concat(/\b/,Y.either(...K.map(ne=>ne.replace(/\s+/,"\\s+"))),/\b/)}const _e={scope:"keyword",match:oe(L),relevance:0};function ce(K,{exceptions:ne,when:ge}={}){const Z=ge;return ne=ne||[],K.map(J=>J.match(/\|\d+$/)||ne.includes(J)?J:Z(J)?`${J}|0`:J)}return{name:"SQL",case_insensitive:!0,illegal:/[{}]|<\//,keywords:{$pattern:/\b[\w\.]+/,keyword:ce(N,{when:K=>K.length<3}),literal:S,type:o,built_in:C},contains:[{scope:"type",match:oe(w)},_e,G,R,U,q,E.C_NUMBER_MODE,E.C_BLOCK_COMMENT_MODE,j,T]}}var he="[0-9](_*[0-9])*",Ae=`\\.(${he})`,Ce="[0-9a-fA-F](_*[0-9a-fA-F])*",ft={className:"number",variants:[{begin:`(\\b(${he})((${Ae})|\\.)?|(${Ae}))[eE][+-]?(${he})[fFdD]?\\b`},{begin:`\\b(${he})((${Ae})[fFdD]?\\b|\\.([fFdD]\\b)?)`},{begin:`(${Ae})[fFdD]?\\b`},{begin:`\\b(${he})[fFdD]\\b`},{begin:`\\b0[xX]((${Ce})\\.?|(${Ce})?\\.(${Ce}))[pP][+-]?(${he})[fFdD]?\\b`},{begin:"\\b(0|[1-9](_*[0-9])*)[lL]?\\b"},{begin:`\\b0[xX](${Ce})[lL]?\\b`},{begin:"\\b0(_*[0-7])*[lL]?\\b"},{begin:"\\b0[bB][01](_*[01])*[lL]?\\b"}],relevance:0};function pt(E,Y,j){return j===-1?"":E.replace(Y,U=>pt(E,Y,j-1))}function nn(E){const Y=E.regex,j="[À-ʸa-zA-Z_$][À-ʸa-zA-Z_$0-9]*",U=j+pt("(?:<"+j+"~~~(?:\\s*,\\s*"+j+"~~~)*>)?",/~~~/g,2),_={keyword:["synchronized","abstract","private","var","static","if","const ","for","while","strictfp","finally","protected","import","native","final","void","enum","else","break","transient","catch","instanceof","volatile","case","assert","package","default","public","try","switch","continue","throws","protected","public","private","module","requires","exports","do","sealed","yield","permits","goto","when"],literal:["false","true","null"],type:["char","boolean","long","float","int","byte","short","double"],built_in:["super","this"]},v={className:"meta",begin:"@"+j,contains:[{begin:/\(/,end:/\)/,contains:["self"]}]},x={className:"params",begin:/\(/,end:/\)/,keywords:_,relevance:0,contains:[E.C_BLOCK_COMMENT_MODE],endsParent:!0};return{name:"Java",aliases:["jsp"],keywords:_,illegal:/<\/|#/,contains:[E.COMMENT("/\\*\\*","\\*/",{relevance:0,contains:[{begin:/\w+@/,relevance:0},{className:"doctag",begin:"@[A-Za-z]+"}]}),{begin:/import java\.[a-z]+\./,keywords:"import",relevance:2},E.C_LINE_COMMENT_MODE,E.C_BLOCK_COMMENT_MODE,{begin:/"""/,end:/"""/,className:"string",contains:[E.BACKSLASH_ESCAPE]},E.APOS_STRING_MODE,E.QUOTE_STRING_MODE,{match:[/\b(?:class|interface|enum|extends|implements|new)/,/\s+/,j],className:{1:"keyword",3:"title.class"}},{match:/non-sealed/,scope:"keyword"},{begin:[Y.concat(/(?!else)/,j),/\s+/,j,/\s+/,/=(?!=)/],className:{1:"type",3:"variable",5:"operator"}},{begin:[/record/,/\s+/,j],className:{1:"keyword",3:"title.class"},contains:[x,E.C_LINE_COMMENT_MODE,E.C_BLOCK_COMMENT_MODE]},{beginKeywords:"new throw return else",relevance:0},{begin:["(?:"+U+"\\s+)",E.UNDERSCORE_IDENT_RE,/\s*(?=\()/],className:{2:"title.function"},keywords:_,contains:[{className:"params",begin:/\(/,end:/\)/,keywords:_,relevance:0,contains:[v,E.APOS_STRING_MODE,E.QUOTE_STRING_MODE,ft,E.C_BLOCK_COMMENT_MODE]},E.C_LINE_COMMENT_MODE,E.C_BLOCK_COMMENT_MODE]},ft,v]}}var ke={exports:{}};/*!
+ * clipboard.js v2.0.11
+ * https://clipboardjs.com/
+ *
+ * Licensed MIT © Zeno Rocha
+ */var rn=ke.exports,gt;function on(){return gt||(gt=1,function(E,Y){(function(U,q){E.exports=q()})(rn,function(){return function(){var j={686:function(S,w,o){o.d(w,{default:function(){return je}});var _=o(279),v=o.n(_),x=o(370),C=o.n(x),L=o(817),F=o.n(L);function N(y){try{return document.execCommand(y)}catch{return!1}}var R=function(u){var c=F()(u);return N("cut"),c},T=R;function G(y){var u=document.documentElement.getAttribute("dir")==="rtl",c=document.createElement("textarea");c.style.fontSize="12pt",c.style.border="0",c.style.padding="0",c.style.margin="0",c.style.position="absolute",c.style[u?"right":"left"]="-9999px";var d=window.pageYOffset||document.documentElement.scrollTop;return c.style.top="".concat(d,"px"),c.setAttribute("readonly",""),c.value=y,c}var oe=function(u,c){var d=G(u);c.container.appendChild(d);var b=F()(d);return N("copy"),d.remove(),b},_e=function(u){var c=arguments.length>1&&arguments[1]!==void 0?arguments[1]:{container:document.body},d="";return typeof u=="string"?d=oe(u,c):u instanceof HTMLInputElement&&!["text","search","url","tel","password"].includes(u==null?void 0:u.type)?d=oe(u.value,c):(d=F()(u),N("copy")),d},ce=_e;function K(y){"@babel/helpers - typeof";return typeof Symbol=="function"&&typeof Symbol.iterator=="symbol"?K=function(c){return typeof c}:K=function(c){return c&&typeof Symbol=="function"&&c.constructor===Symbol&&c!==Symbol.prototype?"symbol":typeof c},K(y)}var ne=function(){var u=arguments.length>0&&arguments[0]!==void 0?arguments[0]:{},c=u.action,d=c===void 0?"copy":c,b=u.container,O=u.target,X=u.text;if(d!=="copy"&&d!=="cut")throw new Error('Invalid "action" value, use either "copy" or "cut"');if(O!==void 0)if(O&&K(O)==="object"&&O.nodeType===1){if(d==="copy"&&O.hasAttribute("disabled"))throw new Error('Invalid "target" attribute. Please use "readonly" instead of "disabled" attribute');if(d==="cut"&&(O.hasAttribute("readonly")||O.hasAttribute("disabled")))throw new Error(`Invalid "target" attribute. You can't cut text from elements with "readonly" or "disabled" attributes`)}else throw new Error('Invalid "target" value, use a valid Element');if(X)return ce(X,{container:b});if(O)return d==="cut"?T(O):ce(O,{container:b})},ge=ne;function Z(y){"@babel/helpers - typeof";return typeof Symbol=="function"&&typeof Symbol.iterator=="symbol"?Z=function(c){return typeof c}:Z=function(c){return c&&typeof Symbol=="function"&&c.constructor===Symbol&&c!==Symbol.prototype?"symbol":typeof c},Z(y)}function J(y,u){if(!(y instanceof u))throw new TypeError("Cannot call a class as a function")}function be(y,u){for(var c=0;c<u.length;c++){var d=u[c];d.enumerable=d.enumerable||!1,d.configurable=!0,"value"in d&&(d.writable=!0),Object.defineProperty(y,d.key,d)}}function xe(y,u,c){return u&&be(y.prototype,u),c&&be(y,c),y}function De(y,u){if(typeof u!="function"&&u!==null)throw new TypeError("Super expression must either be null or a function");y.prototype=Object.create(u&&u.prototype,{constructor:{value:y,writable:!0,configurable:!0}}),u&&ye(y,u)}function ye(y,u){return ye=Object.setPrototypeOf||function(d,b){return d.__proto__=b,d},ye(y,u)}function le(y){var u=Be();return function(){var d=re(y),b;if(u){var O=re(this).constructor;b=Reflect.construct(d,arguments,O)}else b=d.apply(this,arguments);return Le(this,b)}}function Le(y,u){return u&&(Z(u)==="object"||typeof u=="function")?u:Ie(y)}function Ie(y){if(y===void 0)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return y}function Be(){if(typeof Reflect>"u"||!Reflect.construct||Reflect.construct.sham)return!1;if(typeof Proxy=="function")return!0;try{return Date.prototype.toString.call(Reflect.construct(Date,[],function(){})),!0}catch{return!1}}function re(y){return re=Object.setPrototypeOf?Object.getPrototypeOf:function(c){return c.__proto__||Object.getPrototypeOf(c)},re(y)}function Ee(y,u){var c="data-clipboard-".concat(y);if(u.hasAttribute(c))return u.getAttribute(c)}var Pe=function(y){De(c,y);var u=le(c);function c(d,b){var O;return J(this,c),O=u.call(this),O.resolveOptions(b),O.listenClick(d),O}return xe(c,[{key:"resolveOptions",value:function(){var b=arguments.length>0&&arguments[0]!==void 0?arguments[0]:{};this.action=typeof b.action=="function"?b.action:this.defaultAction,this.target=typeof b.target=="function"?b.target:this.defaultTarget,this.text=typeof b.text=="function"?b.text:this.defaultText,this.container=Z(b.container)==="object"?b.container:document.body}},{key:"listenClick",value:function(b){var O=this;this.listener=C()(b,"click",function(X){return O.onClick(X)})}},{key:"onClick",value:function(b){var O=b.delegateTarget||b.currentTarget,X=this.action(O)||"copy",ve=ge({action:X,container:this.container,target:this.target(O),text:this.text(O)});this.emit(ve?"success":"error",{action:X,text:ve,trigger:O,clearSelection:function(){O&&O.focus(),window.getSelection().removeAllRanges()}})}},{key:"defaultAction",value:function(b){return Ee("action",b)}},{key:"defaultTarget",value:function(b){var O=Ee("target",b);if(O)return document.querySelector(O)}},{key:"defaultText",value:function(b){return Ee("text",b)}},{key:"destroy",value:function(){this.listener.destroy()}}],[{key:"copy",value:function(b){var O=arguments.length>1&&arguments[1]!==void 0?arguments[1]:{container:document.body};return ce(b,O)}},{key:"cut",value:function(b){return T(b)}},{key:"isSupported",value:function(){var b=arguments.length>0&&arguments[0]!==void 0?arguments[0]:["copy","cut"],O=typeof b=="string"?[b]:b,X=!!document.queryCommandSupported;return O.forEach(function(ve){X=X&&!!document.queryCommandSupported(ve)}),X}}]),c}(v()),je=Pe},828:function(S){var w=9;if(typeof Element<"u"&&!Element.prototype.matches){var o=Element.prototype;o.matches=o.matchesSelector||o.mozMatchesSelector||o.msMatchesSelector||o.oMatchesSelector||o.webkitMatchesSelector}function _(v,x){for(;v&&v.nodeType!==w;){if(typeof v.matches=="function"&&v.matches(x))return v;v=v.parentNode}}S.exports=_},438:function(S,w,o){var _=o(828);function v(L,F,N,R,T){var G=C.apply(this,arguments);return L.addEventListener(N,G,T),{destroy:function(){L.removeEventListener(N,G,T)}}}function x(L,F,N,R,T){return typeof L.addEventListener=="function"?v.apply(null,arguments):typeof N=="function"?v.bind(null,document).apply(null,arguments):(typeof L=="string"&&(L=document.querySelectorAll(L)),Array.prototype.map.call(L,function(G){return v(G,F,N,R,T)}))}function C(L,F,N,R){return function(T){T.delegateTarget=_(T.target,F),T.delegateTarget&&R.call(L,T)}}S.exports=x},879:function(S,w){w.node=function(o){return o!==void 0&&o instanceof HTMLElement&&o.nodeType===1},w.nodeList=function(o){var _=Object.prototype.toString.call(o);return o!==void 0&&(_==="[object NodeList]"||_==="[object HTMLCollection]")&&"length"in o&&(o.length===0||w.node(o[0]))},w.string=function(o){return typeof o=="string"||o instanceof String},w.fn=function(o){var _=Object.prototype.toString.call(o);return _==="[object Function]"}},370:function(S,w,o){var _=o(879),v=o(438);function x(N,R,T){if(!N&&!R&&!T)throw new Error("Missing required arguments");if(!_.string(R))throw new TypeError("Second argument must be a String");if(!_.fn(T))throw new TypeError("Third argument must be a Function");if(_.node(N))return C(N,R,T);if(_.nodeList(N))return L(N,R,T);if(_.string(N))return F(N,R,T);throw new TypeError("First argument must be a String, HTMLElement, HTMLCollection, or NodeList")}function C(N,R,T){return N.addEventListener(R,T),{destroy:function(){N.removeEventListener(R,T)}}}function L(N,R,T){return Array.prototype.forEach.call(N,function(G){G.addEventListener(R,T)}),{destroy:function(){Array.prototype.forEach.call(N,function(G){G.removeEventListener(R,T)})}}}function F(N,R,T){return v(document.body,N,R,T)}S.exports=x},817:function(S){function w(o){var _;if(o.nodeName==="SELECT")o.focus(),_=o.value;else if(o.nodeName==="INPUT"||o.nodeName==="TEXTAREA"){var v=o.hasAttribute("readonly");v||o.setAttribute("readonly",""),o.select(),o.setSelectionRange(0,o.value.length),v||o.removeAttribute("readonly"),_=o.value}else{o.hasAttribute("contenteditable")&&o.focus();var x=window.getSelection(),C=document.createRange();C.selectNodeContents(o),x.removeAllRanges(),x.addRange(C),_=x.toString()}return _}S.exports=w},279:function(S){function w(){}w.prototype={on:function(o,_,v){var x=this.e||(this.e={});return(x[o]||(x[o]=[])).push({fn:_,ctx:v}),this},once:function(o,_,v){var x=this;function C(){x.off(o,C),_.apply(v,arguments)}return C._=_,this.on(o,C,v)},emit:function(o){var _=[].slice.call(arguments,1),v=((this.e||(this.e={}))[o]||[]).slice(),x=0,C=v.length;for(x;x<C;x++)v[x].fn.apply(v[x].ctx,_);return this},off:function(o,_){var v=this.e||(this.e={}),x=v[o],C=[];if(x&&_)for(var L=0,F=x.length;L<F;L++)x[L].fn!==_&&x[L].fn._!==_&&C.push(x[L]);return C.length?v[o]=C:delete v[o],this}},S.exports=w,S.exports.TinyEmitter=w}},U={};function q(S){if(U[S])return U[S].exports;var w=U[S]={exports:{}};return j[S](w,w.exports,q),w.exports}return function(){q.n=function(S){var w=S&&S.__esModule?function(){return S.default}:function(){return S};return q.d(w,{a:w}),w}}(),function(){q.d=function(S,w){for(var o in w)q.o(w,o)&&!q.o(S,o)&&Object.defineProperty(S,o,{enumerable:!0,get:w[o]})}}(),function(){q.o=function(S,w){return Object.prototype.hasOwnProperty.call(S,w)}}(),q(686)}().default})}(ke)),ke.exports}var sn=on();const an=dt(sn);qe.registerLanguage("sql",tn);qe.registerLanguage("java",nn);qe.highlightAll();function cn(){const E=document.getElementById("copyToast");E.style.display="block",E.classList.add("show"),setTimeout(()=>{E.classList.remove("show"),setTimeout(()=>{E.style.display="none"},150)},2e3)}document.addEventListener("DOMContentLoaded",()=>{new an(".clipboard").on("success",cn)});
