+++ src/main/resources/static/js/designated/dbSetting/columnForm.js
