--- src/main/resources/static/js/script.js
@@ -1,3 +1,0 @@
-/**
- * 
- */
