+++ src/main/resources/templates/contents/databases/detailDB.html
@@ -3,9 +3,9 @@
 	layout:decorate="~{layouts/parentLayout.html}">
 
 <th:block layout:fragment="meta">
-	<script th:src="@{/js/accordion.js}"></script>
-	<script th:src="@{/js/toggleForm.js}"></script>
-	<script th:src="@{/js/dbSetting/columnForm.js}"></script>
+	<script th:src="@{/js/ui/accordion.js}"></script>
+	<script th:src="@{/js/ui/toggleForm.js}"></script>
+	<script th:src="@{/js/designated/dbSetting/columnForm.js}"></script>
 </th:block>
 
 <th:block layout:fragment="breadCrumb">
