+++ src/main/resources/templates/contents/databases/detailTable.html
@@ -5,12 +5,12 @@
 <th:block layout:fragment="meta">
 
 	<!-- これだけでOK：codeblock.js に highlight.js / clipboard.js をすべて含めてある -->
-	<script th:src="@{/js/bundle/codeblock.js}"></script>
-	<link rel="stylesheet" th:href="@{/js/bundle/codeblock.css}" />
+	<script th:src="@{/assets/codeblock/codeblock.js}"></script>
+	<link rel="stylesheet" th:href="@{/assets/codeblock/codeblock.css}" />
 
 
-	<script th:src="@{/js/toggleForm.js}"></script>
-	<script th:src="@{/js/dbSetting/columnForm.js}"></script>
+	<script th:src="@{/js/ui/toggleForm.js}"></script>
+	<script th:src="@{/js/designated/dbSetting/columnForm.js}"></script>
 </th:block>
