+++ src/main/resources/templates/layouts/printLayout.html
@@ -22,7 +22,7 @@
 	<link rel="stylesheet" th:href="@{/webjars/bootstrap/css/bootstrap.min.css}" />
 	<script th:src="@{/webjars/bootstrap/js/bootstrap.min.js}"></script>
 
-	<script src="/js/script.js" th:src="@{/js/script.js}"></script>
+  <link href="https://unpkg.com/boxicons@2.1.4/css/boxicons.min.css" rel="stylesheet">
 
 	<th:block layout:fragment="meta"></th:block>
